app/requirements.txt
Flask
flask-mysqldb
mysql-connector
flask-jwt-extended
Werkzeug
Jinja2
passlib